# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rayzan Mart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (Admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}/block:
    put:
      operationId: toggleUserBlock
      tags: [admin]
      summary: Block/unblock user (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/affiliates/{id}/status:
    put:
      operationId: updateAffiliateStatus
      tags: [admin]
      summary: Update affiliate status (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAffiliateStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/withdrawals/{id}/status:
    put:
      operationId: updateWithdrawalStatus
      tags: [admin]
      summary: Update withdrawal status (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWithdrawalStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/profit-loss:
    get:
      operationId: getProfitLoss
      tags: [admin]
      summary: Get profit/loss report (Admin)
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profit/loss data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfitLossRow"

  # Site Settings
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    UserWithProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
        roles:
          type: array
          items:
            type: string
      required: [id, email]
    Profile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        loyaltyPoints:
          type: integer
        isBlocked:
          type: boolean
        dateOfBirth:
          type: string
          nullable: true
        occupation:
          type: string
          nullable: true
        nid:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        paymentNumber:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, userId, name]
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserWithProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, limit, totalPages]
    BlockUserBody:
      type: object
      properties:
        isBlocked:
          type: boolean
      required: [isBlocked]
    UpdateAffiliateStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, inactive, approved, rejected]
      required: [status]
    UpdateWithdrawalStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        adminNotes:
          type: string
      required: [status]
    ProfitLossRow:
      type: object
      properties:
        orderDate:
          type: string
        categoryName:
          type: string
        totalSales:
          type: number
        totalProductCost:
          type: number
        totalDeliveryCost:
          type: number
        totalCommissions:
          type: number
        netProfit:
          type: number
        orderCount:
          type: integer
      required: [orderDate, categoryName, totalSales, netProfit, orderCount]