# lib/api-zod/src/generated/types/userWithProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rayzan Mart API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Profile } from "./profile";

export interface UserWithProfile {
  id: string;
  email: string;
  profile?: Profile;
  roles?: string[];
}